% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryLevelData.R
\name{countryLevelData}
\alias{countryLevelData}
\title{Global IDP Data Country Level}
\usage{
countryLevelData(
  operation = "",
  countryName = "",
  admin0Pcode = "",
  fromDate = NULL,
  toDate = NULL,
  monthFrom_month = NULL,
  monthFrom_year = NULL,
  monthTo_month = NULL,
  monthTo_year = NULL,
  roundFrom = NULL,
  roundTo = NULL,
  to_dataframe = FALSE
)
}
\arguments{
\item{operation}{String, specifying the operation to perform.}

\item{countryName}{String, specifying the name of country.}

\item{admin0Pcode}{String, specifying the ISO Country Code.}

\item{fromDate}{String, specifying the start date of the reporting period (format: "yyyy-mm-dd"), optional.}

\item{toDate}{String, specifying the start date of the reporting period (format: "yyyy-mm-dd"), optional.}

\item{monthFrom_month}{String, specifying the start month of the reporting period. Between 1 and 12.}

\item{monthFrom_year}{Integer, specifying the start year of the reporting period. Year must be 4 digit.}

\item{monthTo_month}{Integer, specifying the end month of the reporting period. Between 1 and 12.}

\item{monthTo_year}{Integer, specifying the end year of the reporting period. Year must be 4 digit.}

\item{roundFrom}{Integer, specifying the start round number, optional. Not less than 0.}

\item{roundTo}{Integer, specifying the end round number, optional. Not less than 0.}

\item{to_dataframe}{Boolean, Convert the response to a DataFrame, default to False, optional.}
}
\value{
A json file or dataframe.
}
\description{
Global IDP Data Country Level
}
\examples{
countryLevelData(admin0Pcode="MNG", monthFrom_month= "1", monthFrom_year=2000, monthTo_month= "12", monthTo_year=2023,to_dataframe =TRUE)
}
